Fix vimcat deadlock on Linux, avoid exit polling.
**Fix a deadlock affecting Linux under certain circumstances**

When vimcat is run under man as MANPAGER, it will deadlock and hang
after printing the first line of the input for certain manual pages.

For example, `MANPAGER=vimcat man timeout` reliably produces a hang.

When the hang happens, the cat process that read the first line from the
FIFO has exited and is defunct. Its parent process (a shell) is waiting
for a SIGCHLD, suggesting that it is correctly inside a `wait` call, but
the signal never arrives.

Because the shell `wait` never finishes, it never calls cat again to
read the next line from the FIFO, and the vim subprocess also blocks
forever trying to write to the FIFO.

I wasn't able to figure out exactly why this happens, but switching to
job control and using `fg` instead of `wait` allows this to work
reliably and ensure that child processes are correctly reaped.

**Refactor hacky cat/tail to avoid polling, call cat only once**

Previously, vimcat used a subshell that executed a cat subprocess once
for each line of the output. This was done because vim writefile()
closes the file handle to the FIFO after each line written, leading cat
on the other end to believe that EOF has arrived and therefore to exit.

Instead, pipe vim output to /dev/stdout, and have the shell handle
writing to the FIFO. Because the shell won't close its file handle until
vim exits, we can execute cat only once, and it will read the entire
output. This also means we no longer need to kill cat at the end of the
file because it will exit gracefully on its own, and we no longer need
to poll for the vim_done sentinel file because we can simply wait for
vim and cat to exit.